order_queue/requirements.txt
grpcio==1.78.0
grpcio-tools==1.78.0
protobuf==6.33.5
watchdog==6.0.0